+++ common/src/main/java/me/juancarloscp52/spyglass_improvements/mixin/CameraMixin.java
@@ -1,13 +1,13 @@
 package me.juancarloscp52.spyglass_improvements.mixin;
 
 import me.juancarloscp52.spyglass_improvements.client.SpyglassImprovementsClient;
-import net.minecraft.client.renderer.GameRenderer;
+import net.minecraft.client.Camera;
 import org.spongepowered.asm.mixin.Mixin;
 import org.spongepowered.asm.mixin.injection.Constant;
 import org.spongepowered.asm.mixin.injection.ModifyConstant;
 
-@Mixin(GameRenderer.class)
-public class GameRendererMixin {
+@Mixin(Camera.class)
+public class CameraMixin {
 
     @ModifyConstant(method = "tickFov", constant = @Constant(floatValue = 0.1f))
     private float modifyMaximumFov(float constant){