+++ common/src/main/java/me/juancarloscp52/spyglass_improvements/mixin/InGameHudMixin.java
@@ -5,7 +5,7 @@ import me.juancarloscp52.spyglass_improvements.client.SpyglassImprovementsClient
 import net.minecraft.client.DeltaTracker;
 import net.minecraft.client.Minecraft;
 import net.minecraft.client.gui.Gui;
-import net.minecraft.client.gui.GuiGraphics;
+import net.minecraft.client.gui.GuiGraphicsExtractor;
 import net.minecraft.resources.Identifier;
 import org.spongepowered.asm.mixin.Mixin;
 import org.spongepowered.asm.mixin.injection.At;
@@ -17,7 +17,7 @@ import org.spongepowered.asm.mixin.injection.callback.CallbackInfo;
 public class InGameHudMixin {
 
     // Set the spyglass overlay depending on the selected one.
-    @ModifyArg(method = "renderSpyglassOverlay",at = @At(value = "INVOKE",target = "Lnet/minecraft/client/gui/GuiGraphics;blit(Lcom/mojang/blaze3d/pipeline/RenderPipeline;Lnet/minecraft/resources/Identifier;IIFFIIII)V"),index = 1)
+    @ModifyArg(method = "extractSpyglassOverlay",at = @At(value = "INVOKE",target = "Lnet/minecraft/client/gui/GuiGraphicsExtractor;blit(Lcom/mojang/blaze3d/pipeline/RenderPipeline;Lnet/minecraft/resources/Identifier;IIFFIIII)V"),index = 1)
     public Identifier setTexture(Identifier identifier){
         return switch (SpyglassImprovementsClient.getInstance().settings.overlay) {
             case 1 -> Identifier.fromNamespaceAndPath("spyglass_improvements", "textures/spyglass_scope_clear.png");
@@ -26,15 +26,15 @@ public class InGameHudMixin {
         };
     }
     // toggle renderCrosshair depending on settings
-    @Inject(method = "renderCrosshair", at = @At("HEAD"), cancellable = true)
-    public void renderCrosshair(GuiGraphics guiGraphics, DeltaTracker deltaTracker, CallbackInfo ci){
+    @Inject(method = "extractCrosshair", at = @At("HEAD"), cancellable = true)
+    public void renderCrosshair(GuiGraphicsExtractor graphics, DeltaTracker deltaTracker, CallbackInfo ci){
         if(!SpyglassImprovementsClient.getInstance().settings.showCrossHair && Minecraft.getInstance().player!=null && Minecraft.getInstance().player.isScoping())
             ci.cancel();
     }
 
     // Toggle overlay.
-    @WrapWithCondition(method = "Lnet/minecraft/client/gui/Gui;renderCameraOverlays(Lnet/minecraft/client/gui/GuiGraphics;Lnet/minecraft/client/DeltaTracker;)V", at = @At(value = "INVOKE", target = "Lnet/minecraft/client/gui/Gui;renderSpyglassOverlay(Lnet/minecraft/client/gui/GuiGraphics;F)V"))
-    public boolean DoNotRenderIfNoneOverlay(Gui instance, GuiGraphics guiGraphics, float f){ // No overlay.
+    @WrapWithCondition(method = "extractCameraOverlays", at = @At(value = "INVOKE", target = "Lnet/minecraft/client/gui/Gui;extractSpyglassOverlay(Lnet/minecraft/client/gui/GuiGraphicsExtractor;F)V"))
+    public boolean DoNotRenderIfNoneOverlay(Gui instance, GuiGraphicsExtractor graphics, float scale){ // No overlay.
         return SpyglassImprovementsClient.getInstance().settings.overlay != 3;
     }