+++ fabric/src/main/java/me/juancarloscp52/spyglass_improvements/fabric/client/SpyglassImprovementsFabricClient.java
@@ -6,14 +6,14 @@ import me.juancarloscp52.spyglass_improvements.fabric.client.integrations.Trinke
 import me.juancarloscp52.spyglass_improvements.integrations.AccessoriesIntegration;
 import net.fabricmc.api.ClientModInitializer;
 import net.fabricmc.fabric.api.client.event.lifecycle.v1.ClientTickEvents;
-import net.fabricmc.fabric.api.client.keybinding.v1.KeyBindingHelper;
+import net.fabricmc.fabric.api.client.keymapping.v1.KeyMappingHelper;
 import net.fabricmc.loader.api.FabricLoader;
 
 public final class SpyglassImprovementsFabricClient implements ClientModInitializer {
 
     @Override
     public void onInitializeClient() {
-        KeyBindingHelper.registerKeyBinding(SpyglassImprovementsClient.useSpyglass);
+        KeyMappingHelper.registerKeyMapping(SpyglassImprovementsClient.useSpyglass);
         IEquipmentIntegration integration = null;
         if(FabricLoader.getInstance().isModLoaded("accessories")){
             integration = new AccessoriesIntegration();